Rename wcecoli_state to json_state
