+++ ecoli/experiments/tet_amp_sim.py
@@ -9,7 +9,7 @@ from ecoli.experiments.ecoli_engine_process import run_simulation
 from ecoli.experiments.ecoli_master_sim import CONFIG_DIR_PATH, SimConfig
 from ecoli.library.parameters import param_store
 from ecoli.library.logging_tools import write_json
-from ecoli.library.wcecoli_state import get_state_from_file
+from ecoli.library.json_state import get_state_from_file
 
 AVOGADRO = constants.N_A / units.mol
